requirements.txt
ultralytics
fastapi
uvicorn
opencv-python
numpy
pandas
sqlalchemy
psycopg2-binary
